refactor(App): 기능 파일 통합
getBall, getStrike, createResult, checkPlayingNum, checkRestartNum, checkException, restartQuestion, getUserInput, createAnswer, startGame, play 기능 class App 내로 통합
