--- src/createAnswer.js
@@ -1,14 +1,0 @@
-const MissionUtils = require("@woowacourse/mission-utils");
-
-const createAnswer = () => {
-  const computer = [];
-  while (computer.length < 3) {
-    const number = MissionUtils.Random.pickNumberInRange(1, 9);
-    if (!computer.includes(number)) {
-      computer.push(number);
-    }
-  }
-  return computer.join("");
-};
-
-module.exports = createAnswer;
