--- src/getUserInput.js
@@ -1,22 +1,0 @@
-const MissionUtils = require("@woowacourse/mission-utils");
-const Message = require("./message/message");
-const checkException = require("./checkException");
-const $utils = MissionUtils.Console;
-
-const getUserInput = () => {
-  const playingInput = 0;
-  let userInput = "";
-  $utils.readLine(Message.INPUT, (input) => {
-    $utils.print(input);
-    if (!checkException(input, playingInput)) {
-      $utils.print(Message.ERROR);
-      throw Message.ERROR;
-    }
-    userInput = input;
-    $utils.close();
-  });
-
-  return userInput;
-};
-
-module.exports = getUserInput;
