--- src/message/message.js
@@ -1,10 +1,0 @@
-const Message = {
-  START: "숫자 야구 게임을 시작합니다.",
-  INPUT: "숫자를 입력해주세요 : ",
-  ERROR: "잘못된 문자를 입력하였습니다. 프로그램을 종료합니다.",
-  FINISH: "3개의 숫자를 모두 맞히셨습니다! 게임 종료",
-  CORRECT: "3스트라이크",
-  END: "게임 종료",
-};
-
-module.exports = Message;
